[Refactor] Move Shell Class to header file
