+++ TestShell/TestShell.vcxproj.filters
@@ -56,5 +56,8 @@
     <ClInclude Include="CmdLancher.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
+    <ClInclude Include="Shell.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
