+++ TestShell/main.cpp
@@ -1,7 +1,7 @@
 
 #include "gmock/gmock.h"
 #include <iostream>
-#include "Shell.cpp"
+#include "Shell.h"
 using namespace testing;
